Fix up CCT math, switch to modified Tanner Helland method
This fixes a math error in my implementation of the Brude Lindbloom method.
Since the temp range of that method does not go low enough for my uses,
doesn't offer the color range I hoped for, and I'm fairly confident I'm
using it wrong, this change also switches to the Tanner Helland method.