Set up OTA upgrades
This change adds the ability to run an OTA upgrade via HTTP.

It also includes a script with convenient PowerShell snippets for doing
so.

Please be under no illusion that this just worked--I spent the better
part of a day working on it and hit many bugs. Rebase is my friend.